cleanup and bug fixes
git-svn-id: svn://orca.physics.unc.edu/Orca@11322 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/RunningAverage/ORRunningAverage.h
@@ -30,7 +30,7 @@
 @interface ORRunningAverage : NSObject
 {
     float           runningAverage;
-    unsigned long   dataCount;
+    int             dataCount;
     float           spikeValue;
     float           averageAtTimeOfSpike;
     int             windowLength;
